# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /assistant:
    post:
      operationId: askOperationsAssistant
      tags: [assistant]
      summary: Understand and execute a natural-language store request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssistantRequest"
      responses:
        "200":
          description: Assistant response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistantResponse"
  /assistant/transcribe:
    post:
      operationId: transcribeAssistantVoice
      tags: [assistant]
      summary: Transcribe a recorded manager voice request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TranscriptionRequest"
      responses:
        "200":
          description: Transcript
          content:
            application/json:
              schema:
                type: object
                required: [transcript]
                properties:
                  transcript: { type: string }
        "400": { description: Invalid audio payload }
        "413": { description: Recording too large }
        "415": { description: Unsupported audio format }
        "429": { description: Too many transcription attempts }
        "502": { description: Speech provider could not transcribe the recording }
        "503": { description: Speech service is not configured }
...
components:
  schemas:
...
    AssistantTask:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        area: { type: string }
        due: { type: string }
        done: { type: boolean }
        priority: { type: string, enum: [high, normal] }
        assignee: { type: string }
      required: [id, title, area, due, done]
    AssistantIssue:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        area: { type: string }
        status: { type: string, enum: [Open, In progress, Resolved] }
        time: { type: string }
        urgent: { type: boolean }
      required: [id, title, area, status, time]
    AssistantRequest:
      type: object
      properties:
        request: { type: string, minLength: 1 }
        confirmed: { type: boolean }
        tasks:
          type: array
          items: { $ref: "#/components/schemas/AssistantTask" }
        issues:
          type: array
          items: { $ref: "#/components/schemas/AssistantIssue" }
      required: [request]
    AssistantResponse:
      type: object
      properties:
        message: { type: string }
        intent: { type: string, enum: [answer, create_task, summarize_issues, operations_summary, sop_guidance, clarify] }
        confirmationRequired: { type: boolean }
        task:
          anyOf:
            - { $ref: "#/components/schemas/AssistantTask" }
            - { type: "null" }
        mentions:
          type: array
          items: { type: string }
        auditId: { type: string }
      required: [message, intent, confirmationRequired, mentions, auditId]
    TranscriptionRequest:
      type: object
      required: [audioBase64]
      properties:
        audioBase64: { type: string, format: byte, maxLength: 12000000 }
        mimeType:
          type: string
          description: MIME type of the recorded audio
          enum: [audio/aac, audio/flac, audio/m4a, audio/mp4, audio/mpeg, audio/mp3, audio/ogg, audio/wav, audio/x-wav, audio/webm]
...