# lib/api-zod/src/generated/types/assistantRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AssistantIssue } from './assistantIssue';
import type { AssistantTask } from './assistantTask';

export interface AssistantRequest {
  /** @minLength 1 */
  request: string;
  confirmed?: boolean;
  tasks?: AssistantTask[];
  issues?: AssistantIssue[];
}